+++ unittests/v1.2.0/test_report_runner.py
@@ -1,49 +1,31 @@
-import unittest
-from parameterized import parameterized
-from compliance_suite.report_runner import add_access_methods_test_case
-from unittest.mock import patch, Mock
+from compliance_suite.report_runner import *
 
-class TestReportRunner(unittest.TestCase):
+def test_report_runner():
 
-    @parameterized.expand([
-        ("has_access_methods"),
-        ("has_access_info")])
-    @patch('compliance_suite.report_runner.ValidateDRSObjectResponse')
-    def test_add_access_methods_test_case(self, case_type, MockValidateDRSObjectResponse):
+    actual_final_json = json.loads(report_runner(server_base_url = "http://localhost:8089/ga4gh/drs/v1",
+                                    platform_name = "good mock server",
+                                    platform_description = "test",
+                                    auth_type = "basic"))
 
-        test_object = Mock()
-        case_description = "Test case for access methods"
-        endpoint_name = "Test endpoint"
-        response = Mock()
-        mock_validate_drs_response = Mock()
-        MockValidateDRSObjectResponse.return_value = mock_validate_drs_response
-        skip_access_methods_test_cases = False
-        skip_message=""
-        is_bundle = False
-        access_id_list = add_access_methods_test_case(
-            test_object,
-            case_type,
-            case_description,
-            endpoint_name,
-            response,
-            skip_access_methods_test_cases,
-            skip_message,
-            is_bundle
-        )
+    # TODO: None and BearerAuth 
 
-        # Assertions
-        test_object.add_case.assert_called()
-        test_case = test_object.add_case.return_value
-        test_case.set_case_name.assert_called_with(f"{endpoint_name} has access information")
-        test_case.set_case_description.assert_called_with(case_description)
-        mock_validate_drs_response.set_case.assert_called_with(test_case)
-        mock_validate_drs_response.set_actual_response.assert_called_with(response)
-        test_case.set_end_time_now.assert_called()
+    # remove timestamps, otherwise assert will fail 100%
+    actual_final_json["start_time"] = ""
+    actual_final_json["end_time"] = ""
+    for phase in actual_final_json["phases"]:
+        phase["start_time"] = ""
+        phase["end_time"] = ""
+        for test in phase["tests"]:
+            test["start_time"] = ""
+            test["end_time"] = ""
+            for case in test["cases"]:
+                case["start_time"] = ""
+                case["end_time"] = ""
+            
+    expect_final_json = json.loads(
+        open("unittests/output/expected_good.json", "r").read()
+    )
+    actual_json_s = str(actual_final_json).replace("'", '"').replace("\\","")
+    expect_json_s = str(expect_final_json).replace("'", '"').replace("\\","")
 
-        if case_type == "has_access_methods":
-            mock_validate_drs_response.validate_has_access_methods.assert_called()
-            mock_validate_drs_response.validate_has_access_info.assert_not_called()
-            self.assertIsNone(access_id_list)
-        else:
-            mock_validate_drs_response.validate_has_access_methods.assert_not_called()
-            mock_validate_drs_response.validate_has_access_info.assert_called()
+    assert actual_json_s == expect_json_s
